Moved textFormat into /js/common/utils folder. Moved removeTwitterNameFromDescription into its own file in common folder.
