+++ src/js/components/Organization/OrganizationDisplayForList.jsx
@@ -3,9 +3,9 @@ import React, { Component, Suspense } from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import { renderLog } from '../../common/utils/logging';
+import removeTwitterNameFromDescription from '../../common/utils/removeTwitterNameFromDescription';
 import OrganizationStore from '../../stores/OrganizationStore';
 import { NameAndTwitter, OrganizationDescriptionText, OrganizationDetailsWrapper, OrganizationImage, OrganizationFollowWrapper, OrganizationName, OrganizationLogoWrapper, TwitterOuterWrapper } from '../Style/DisplayForList';
-import { removeTwitterNameFromDescription } from '../../utils/textFormat';
 import PositionInformationOnlySnippet from '../Widgets/PositionInformationOnlySnippet';
 import PositionRatingSnippet from '../Widgets/PositionRatingSnippet';
 import PositionSupportOpposeSnippet from '../Widgets/PositionSupportOpposeSnippet';
