+++ src/js/components/VoterGuide/OrganizationCard.jsx
@@ -7,8 +7,8 @@ import OrganizationActions from '../../actions/OrganizationActions';
 import LoadingWheel from '../../common/components/Widgets/LoadingWheel';
 import { renderLog } from '../../common/utils/logging';
 import stringContains from '../../common/utils/stringContains';
+import removeTwitterNameFromDescription from '../../common/utils/removeTwitterNameFromDescription';
 import OrganizationStore from '../../stores/OrganizationStore';
-import { removeTwitterNameFromDescription } from '../../utils/textFormat';
 import ParsedTwitterDescription from '../Twitter/ParsedTwitterDescription';
 import IssuesByOrganizationDisplayList from '../Values/IssuesByOrganizationDisplayList';
 import PositionInformationOnlySnippet from '../Widgets/PositionInformationOnlySnippet';
