+++ src/js/components/Widgets/PositionSupportOpposeSnippet.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import React, { Component, Suspense } from 'react';
 import normalizedImagePath from '../../common/utils/normalizedImagePath';
 import { renderLog } from '../../common/utils/logging';
-import { vimeoRegX, youTubeRegX } from '../../utils/textFormat';
+import { vimeoRegX, youTubeRegX } from '../../common/utils/textFormat';
 
 const ReadMore = React.lazy(() => import(/* webpackChunkName: 'ReadMore' */ '../../common/components/Widgets/ReadMore'));
 const ReactPlayer = React.lazy(() => import(/* webpackChunkName: 'ReactPlayer' */ 'react-player'));
