+++ src/js/pages/Settings/SettingsDashboard.jsx
@@ -33,7 +33,7 @@ import BallotStore from '../../stores/BallotStore';
 import OrganizationStore from '../../stores/OrganizationStore';
 import VoterGuideStore from '../../stores/VoterGuideStore';
 import VoterStore from '../../stores/VoterStore';
-import { isProperlyFormattedVoterGuideWeVoteId } from '../../utils/textFormat';
+import { isProperlyFormattedVoterGuideWeVoteId } from '../../common/utils/textFormat';
 
 const SignInOptionsPanel = React.lazy(() => import(/* webpackChunkName: 'SignInOptionsPanel' */ '../../components/SignIn/SignInOptionsPanel'));
