+++ src/js/stores/MeasureStore.js
@@ -2,7 +2,7 @@ import { ReduceStore } from 'flux/utils';
 import Dispatcher from '../common/dispatcher/Dispatcher';
 import stringContains from '../common/utils/stringContains';
 import { extractNumberOfPositionsFromPositionList } from '../utils/positionFunctions'; // eslint-disable-line import/no-cycle
-import { stripHtmlFromString } from '../utils/textFormat';
+import { stripHtmlFromString } from '../common/utils/textFormat';
 
 class MeasureStore extends ReduceStore {
   getInitialState () {
